Update readme.txt
Hello, and first of all thank you again for your incredible work. I'm addressing the team to offer you a feedback that I hope will be constructive on the difference between the repo version and the core version.
I've been seeing for several weeks now that the reviews on the extension have been deteriorating strongly. I have read the reviews and interviewed users, friends and colleagues. I realize that in fact the problem probably comes from a misunderstanding of the role of the repo extension compared to the core version of Gutenberg for WordPress.
For a lot of users, the presence of the Gutenberg extension on the repo suggests that it is an extended version of Gutenberg . A bit like if the core version was a lite version. This creates an extra expectation.
Of course, new settings or new features may appear in the extension which justifies this feeling.
Especially since the common strategy of many extension developers or editors to often put a standard version and a premium version can, I think, cause this confusion.
So I suggest that you make it clear in the readme (first place in the introduction) that the repo version is a development (or beta) version and that its use is reserved for informed users to test new features. Thanks